fix(charts): improve sql compatibility
