+++ packages/plugins/charts/src/server/handlers/index.ts
@@ -1,0 +1,5 @@
+export * from './pieHandler';
+export * from './barHandler';
+export * from './columnHandler';
+export * from './lineHandler';
+export * from './areaHandler';
